[#6] main.go: Refactor functions into separate 'parser' package
